IntelliJ version check on enabling, 3 minute cooldown. (#601)
IntelliJ version check on enabling instead of on project start. Don't check if last check was less than 3 minutes ago.
