+++ intellij-ext/src/main/resources/META-INF/plugin.xml
@@ -21,8 +21,6 @@
     <projectListeners>
         <listener class="com.metalbear.mirrord.MirrordListener"
                   topic="com.intellij.execution.ExecutionListener"/>
-        <listener class="com.metalbear.mirrord.VersionCheck"
-                  topic="com.intellij.openapi.project.ProjectManagerListener"/>
     </projectListeners>
 
     <actions>
